Master dev 0.1.3 (#47)
* add noise wrapper

* finish config for doc

* fix bugs; update save load function API; update examples

* add a bash for docs

* update model-based algo

* update

* working on docs

* working on gp

* release v0.1.3

* update doc

+++ baconian/common/spaces/discrete.py
@@ -69,7 +69,7 @@ class Discrete(Space):
         raise NotImplementedError
 
     def clip(self, x):
-        x = np.asarray(x)
+        x = np.asarray(x).astype(np.int)
         assert x.shape == ()
         if self.contains(x):
             return x
